requirements.txt
fastapi==0.61.1
tortoise-orm==0.16.17
uvicorn==0.11.8
aiomysql==0.0.20
aioredis==1.3.1
qiniu==7.6.0
ronglian_sms_sdk==1.0.0
python-alipay-sdk==3.0.4
pyjwt==2.3.0
loguru==0.6.0